# swagger.yml
openapi: 3.0.1
info:
  title: E-Commerce API Documentation
  description: Simple E-Commerce API application made with Express and PostgreSQL.
  version: 1.0.0
servers:
  - url: //localhost:3000/
paths:
  /auth/register:
    summary: Register a new user
    post:
      summary: Registers a new user
      description: Create a user object and adds it to the database
      operationId: register
      responses:
        '200':
          description: 201 Created
        '500':
          description: 500 Internal Server Error
      tags:
        - auth
  /auth/login:
    summary: Logs in a user
    post:
      summary: Logs in a user
      description: >-
        Checks the database for the user and returns successful if they are
        found
      operationId: login
      responses:
        '200':
          description: 200 OK
        '401':
          description: 401 Unauthorized
      tags:
        - auth
  /auth/status:
    summary: Checks a user is logged in
    get:
      summary: Checks if user is logged in
      description: Returns a success
      operationId: loginStatus
      responses:
        '200':
          description: 200 OK
        '401':
          description: 401 Unauthorized
      tags:
        - auth
  /auth/logout:
    summary: Logs out user
    description: Logs out currently logged in user
    post:
      summary: Logs out user
      description: Logs out the current users session
      operationId: logout
      responses:
        '200':
          description: 200 OK
        '400':
          description: 400 Bad Request
        '401':
          description: 401 Unauthorized
      tags:
        - auth
...